Create the plugit_proxy library, fusion code from others plugit instance, update plugit proxy. Use statics for plugit proxy templates
