+++ examples/standalone_proxy/app/settings.py
@@ -83,6 +83,7 @@ STATICFILES_DIRS = (
     # Put strings here, like "/home/html/static" or "C:/www/django/static".
     # Always use forward slashes, even on Windows.
     # Don't forget to use absolute paths, not relative paths.
+    normpath(join(DJANGO_ROOT, 'static')) + '/',
 )
 
 # List of finder classes that know how to find static files in
