+++ examples/standalone_proxy/plugIt/templates/plugIt/api.html
@@ -50,7 +50,7 @@
 
     Returns a URL to the topic or the error.
 
-        <h3>ebuio/forum/search/bytag/&lt;tag&gt;?u=&lt;userPk&gt;'</h3>
+        <h3>/ebuio/forum/search/bytag/&lt;tag&gt;?u=&lt;userPk&gt;</h3>
         <small><i>PlugItAPI.api_ebuio_forum_get_topics_by_tag_for_user()</i></small><br />
 
     Returns a list of all the topics in the application forum with the tag passed in the query.
