+++ examples/standalone_proxy/plugIt/templates/plugIt/base.html
