+++ examples/standalone_proxy/plugIt/templates/plugIt/plugItBase.html
@@ -1,0 +1,1 @@
+{% extends "plugIt/plugItBase-common.html" %}
