+++ examples/standalone_proxy/static/css/bootstrap-theme.css
