+++ examples/standalone_proxy/static/css/bootstrap.css
