+++ examples/standalone_proxy/static/css/plugIt-standalone.css
