+++ examples/standalone_proxy/static/fonts/glyphicons-halflings-regular.woff2
