+++ examples/standalone_proxy/static/fonts/ionicons.woff
