+++ examples/standalone_proxy/static/js/bootstrap.js
