+++ examples/standalone_proxy/static/js/npm.js
