--- examples/standalone_proxy/templates/plugIt/plugItBase.html
@@ -1,137 +1,0 @@
-{% load i18n %}
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <title>{% block title %}{% endblock %} :: PlugIt -Standalone mode</title>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-        {% block css %}
-        {#  CSS used on a specific page. #}
-        {#  use {{ block.super }} to avoid overwriting #}
-        <link href="/media/css/bootstrap.min.css" rel="stylesheet" media="screen">
-        <link href="/media/css/ionicons.min.css" rel="stylesheet" media="screen">
-        <link href="/media/css/plugIt-standalone.css" rel="stylesheet" media="screen">
-        {% endblock %}
-
-        {% block javascript %}
-        {#  Javascript used on a specific page. #}
-        {#  use {{ block.super }} to avoid overwriting #}
-        <script src="/media/js/jquery-2.1.4.min.js"></script>
-        <script src="/media/js/bootstrap.min.js"></script>
-        {% endblock %}
-
-    </head>
-    <body>
-
-        <script type="text/javascript">
-            function ebuio_setUser(mode) {
-                $.ajax({
-                    url: "{{context.ebuio_baseUrl}}ebuio_setUser?mode=" + mode,
-                }).done(function () {
-                    location.reload();
-                });
-            }
-
-            function ebuio_setOrga(name, pk) {
-                $.ajax({
-                    url: "{{context.ebuio_baseUrl}}ebuio_setOrga?name=" + name + "&pk=" + pk,
-                }).done(function () {
-                    location.reload();
-                });
-            }
-        </script>
-
-        <div class="container">
-        <nav class="navbar navbar-inverse">
-            <div class="container-fluid">
-
-                <div class="navbar-header">
-                    <button type="button" class="navbar-toggle collapsed" data-toggle="collapse"
-                            data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
-                        <span class="sr-only">Toggle navigation</span>
-                        <span class="icon-bar"></span>
-                        <span class="icon-bar"></span>
-                        <span class="icon-bar"></span>
-                    </button>
-                    <a class="navbar-brand" href="{{ context.ebuio_baseUrl }}">PlugIt - Standalone mode</a>
-                </div>
-
-                <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-                    <ul class="nav navbar-nav">
-                        {% if context.ebuio_realUsers %}
-
-                            <li>
-                                {% if ebuio_u.username %}
-                                    <a href="{% url 'django.contrib.auth.views.logout' %}"> Welcome {{ ebuio_u.username }}. Logout ?</a>
-                                {% else %}
-                                    <a href="{% url 'django.contrib.auth.views.login' %}"> Not logged. Login ?</a>
-                                {% endif %}
-                                </a>
-
-                            </li>
-
-                            <li>
-                                <a href="/admin/">Admin</a>
-                            </li>
-
-                        {% else %}
-
-                            <li class="dropdown">
-                                <a href="#" class="dropdown-toggle" data-toggle="dropdown">
-                                    {% trans "User" %}
-                                    <b class="caret"></b>
-                                </a>
-                                <ul class="dropdown-menu">
-                                    <li {% if context.ebuio_userMode == 'ano' %}class="active"{% endif %}><a href="#" onclick="ebuio_setUser('ano');">{% trans "Anonymous user" %}</a></li>
-                                    <li {% if context.ebuio_userMode == 'log' %}class="active"{% endif %}><a href="#" onclick="ebuio_setUser('log');">{% trans "Logged in user" %}</a></li>
-                                    <li {% if context.ebuio_userMode == 'mem' %}class="active"{% endif %}><a href="#" onclick="ebuio_setUser('mem');">{% trans "Project member" %}</a></li>
-                                    <li {% if context.ebuio_userMode == 'adm' %}class="active"{% endif %}><a href="#" onclick="ebuio_setUser('adm');">{% trans "Admin user" %}</a></li>
-                                </ul>
-                            </li>
-
-                        {% endif %}
-
-                        {% if context.ebuio_orgamode %}
-                            <li class="dropdown">
-                                <a href="#" class="dropdown-toggle" data-toggle="dropdown">
-                                    {% trans "Organization" %}
-                                    <b class="caret"></b>
-                                </a>
-                                <ul class="dropdown-menu">
-                                    <li {% if context.ebuio_orga.pk == '-1' %}class="active"{% endif %}><a href="#" onclick="ebuio_setOrga('EBU', '-1');">{% trans "EBU" %}</a></li>
-                                    <li {% if context.ebuio_orga.pk == '-2' %}class="active"{% endif %}><a href="#" onclick="ebuio_setOrga('RTS', '-2');">{% trans "RTS" %}</a></li>
-                                    <li {% if context.ebuio_orga.pk == '-3' %}class="active"{% endif %}><a href="#" onclick="ebuio_setOrga('BBC', '-3');">{% trans "BBC" %}</a></li>
-                                    <li {% if context.ebuio_orga.pk == '-4' %}class="active"{% endif %}><a href="#" onclick="ebuio_setOrga('CNN', '-4');">{% trans "CNN" %}</a></li>
-                                </ul>
-                            </li>
-                        {% endif %}
-                    </ul>
-                    <ul class="nav navbar-nav navbar-right">
-                        <li><a href="http://ebu.io" class="text-info">About EBU.io</a></li>
-                        <li><a href="https://github.com/ebu/PlugIt" class="text-info">About PlugIt</a></li>
-                    </ul>
-
-                </div>
-            </div>
-        </nav>
-
-        <div class="row">
-            <div class="col-md-12">
-            {% block content %}{{plugit_content|safe}}
-            {% endblock %}
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-md-12">
-                <hr />
-                <div class="footer">
-                    <p>&copy; EBU 2013 - 2017</p>
-                </div>
-            </div>
-        </div>
-        </div>
-    </body>
-</html>
