+++ plugit/routes.py
@@ -2,7 +2,7 @@ from params import PI_BASE_URL, PI_API_VERSION, PI_API_NAME
 from views import request, check_ip, jsonify, MetaView, TemplateView, ActionView
 
 
-def load_routes(app, actions, mail_callback):
+def load_routes(app, actions, mail_callback=None):
 
     @app.route(PI_BASE_URL + "ping")
     def ping():
@@ -21,6 +21,9 @@ def load_routes(app, actions, mail_callback):
     @app.route(PI_BASE_URL + "mail", methods=['POST'])
     def mail():
         """The mail method: Process mail handling"""
+
+        if not request.form['response_id']:
+            return jsonify(result='Error')
 
         if not mail_callback:
             return jsonify(result='Ok')
@@ -42,7 +45,7 @@ def load_routes(app, actions, mail_callback):
 
             # Template
             app.add_url_rule(
-                '{}tempalte{}'.format(PI_BASE_URL, obj.pi_api_route),
+                '{}template{}'.format(PI_BASE_URL, obj.pi_api_route),
                 view_func=TemplateView.as_view('template_{}'.format(act), action=obj))
 
             # Action
